Extract CollapsableList component.
